chore: backports to v2.x (#70)
* Adds support for ArrayBuffer, Uint8Array and Buffer to send(body) (#45)

* Adds support for ArrayBuffer, Uint8Array and Buffer to send(body)

XMLHttpRequet supports BufferSource as a send parameter, and this is used to transmit binary to the server as part of the POST method.

Node also supports transmitting binary in its HttpClient, which is the underlying implementation in w3c-xmlhttprequest. To utilise this you pass a Buffer into the client.write() method, but w3c-xmlhttprequest only writes String or FormData.

The first patch modifies the check to support Buffer as a body. The second path detect ArrayBuffer or Uint8Array as a body, and then converts to Buffer using Buffer.from(). This means that you can now use the XHR in-place of XMLHttpRequests which pass ArrayBuffer or Uint8Array.

Being able to use Buffer as well in Node is a pleasant side-effect.

Improvements: Buffer.from() actually supports many more types, which could be used. Also, it is possible to convert typed-arrays to Buffer directly.

* delint

* Fix syntax

Co-authored-by: Stuart Espey <stuart.espey@mactrix.com>
Co-authored-by: Yamagishi Kazutoshi <ykzts@desire.sh>

* Add content length (#46)

* Adds support for ArrayBuffer, Uint8Array and Buffer to send(body)

XMLHttpRequet supports BufferSource as a send parameter, and this is used to transmit binary to the server as part of the POST method.

Node also supports transmitting binary in its HttpClient, which is the underlying implementation in w3c-xmlhttprequest. To utilise this you pass a Buffer into the client.write() method, but w3c-xmlhttprequest only writes String or FormData.

The first patch modifies the check to support Buffer as a body. The second path detect ArrayBuffer or Uint8Array as a body, and then converts to Buffer using Buffer.from(). This means that you can now use the XHR in-place of XMLHttpRequests which pass ArrayBuffer or Uint8Array.

Being able to use Buffer as well in Node is a pleasant side-effect.

Improvements: Buffer.from() actually supports many more types, which could be used. Also, it is possible to convert typed-arrays to Buffer directly.

* add Content-Length to requests with a body
Some servers do not accept unbounded requests. Add the Content-Length header to rectify

* Fix lint error
```
/Users/stux/git/node-xmlhttprequest/lib/xmlhttprequest.js
  342:7  error  Keyword "if" must be followed by whitespace  space-after-keywords

✖ 1 problem (1 error, 0 warnings)
```

Fixed.

Co-authored-by: Stuart Espey <stuart.espey@mactrix.com>
Co-authored-by: Yamagishi Kazutoshi <ykzts@desire.sh>

Co-authored-by: Stuart Espey <stux@3ivx.com>
Co-authored-by: Stuart Espey <stuart.espey@mactrix.com>